empirical_evaluation/requirements.txt
# ########## #
# Base       #
# Python 3.7 #
# ########## #
pip==21.0
wheel==0.38.0
setuptools==65.5.0

# ######### #
# Libraries #
# ######### #
networkx==2.6.3
matplotlib==3.5.3
numpy==1.21.6
gym==0.21.0
importlib-metadata==4.13.0
opencv-python==4.9.0.80
pygame==2.5.2
IPython==7.34.0
ipywidgets==8.1.2
scipy==1.7.3
pyyaml==6.0.1
pandas==1.3.5
seaborn==0.12.2
tqdm
